ai_services/requirements.txt
google-genai>=0.3.0
pydantic>=2.0.0
PyYAML>=6.0
fastapi>=0.111.0
uvicorn>=0.30.0
psycopg2-binary>=2.9.0
pymongo>=4.6.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
langchain>=0.2.0
langchain-community>=0.2.0
pypdf>=4.0.0
docx2txt>=0.8
pytest>=7.0.0
pytest-asyncio>=0.21.0
